put the integration tests with the github duck tests
